// packages/discord/src/operations/getOriginalWebhookMessage.ts
export const GetOriginalWebhookMessageInput =
  /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
    webhook_id: Schema.String.pipe(T.PathParam()),
    webhook_token: Schema.String.pipe(T.PathParam()),
    thread_id: Schema.optional(Schema.String),
  }).pipe(
    T.Http({
      method: "GET",
      path: "/webhooks/{webhook_id}/{webhook_token}/messages/@original",
    }),
  );
export type GetOriginalWebhookMessageInput =
  typeof GetOriginalWebhookMessageInput.Type;
// ...
export type GetOriginalWebhookMessageOutput =
  typeof GetOriginalWebhookMessageOutput.Type;

// The operation
export const getOriginalWebhookMessage = /*@__PURE__*/ /*#__PURE__*/ API.make(
  () => ({
    inputSchema: GetOriginalWebhookMessageInput,
    outputSchema: GetOriginalWebhookMessageOutput,
    errors: [BadRequest, Forbidden, NotFound] as const,
  }),
);